Cursor now takes position of a block in the list.
Figured out how to get the index of a linear list from a 2D equivalent index. Next I need to make the cursor capture the blocks that it's resting on so they can be switched. Also, need to increase the dead zone on the left stick for better cursor movement.