Fixed block swapping
The problem was that when I created a temp block and assigned it to an object from the block list, I was only creating a reference to the position in the list and not a new copy of it. So whenever the position in the list changed, it would change the temp block's position.