Vertical match half working
I figured out the out of range error. The problem was the y iterator not pointing to correct value in the match list, which eventually caused y to point to a null object. The match list still isn't returning a correct set of matching blocks, but progress was made regardless!