Vertical matching fixed!
My logic for checking the last 2 blocks was incorrect. I didn't include the possibility that the last 2 blocks could be the same colour, which could've meant those blocks weren't part of a match, or they were. The solution I found was to look at their block IDs, and compare them with the ID list.