Added row and col value to blocks
After getting some help on Discord, it turns out I was overthinking things again. To make a block fall, I only need to check for any blocks in the location below the block that needs to fall. If there's none, then just set the block to the new location. I added a row and block value to track the block's location in the well. (Should've done this from the start)