Matching issues
Blocks can no longer be matched after a certain number of matches are made. Might have to do with the bottom row, which should not be included in matching if it's not completely past the red line. When matching stops working, blocks will not fall and the cursor is no longer functional.