Fixed issue with block matching. Previous fix to cursor not moving past certain point
The problem was the new row depth code I wrote to try to exclude the empty rows. I want to eliminate the empty rows so the block list doesn't get too large, but don't have a good solution yet.

Also, I managed to fix the cursor problem but didn't mention it in the previous commit. The cursor will now always move to the bottom of the well.