**IMPORTANT COMMIT** Adding 2D array for cursor to move around in
This commit is critical because it's the start of a major change I'm making to the game. I'm adding a 2D array for the cursor to move around in. It is the "map". I'm still using a list for the blocks, but I expect much of the code will have to change from this point.